server/requirements.txt
# Install with: `pip install -r requirements.txt`

Django==1.10.5
psycopg2==2.6.2
django-extensions==1.7.6

# Admin
django-ckeditor==5.2.1
django-autocomplete-light==3.2.1

# Rest Framework
djangorestframework==3.5.4
Markdown==2.6.8
django-filter==1.0.1

# CORS
django-cors-headers==2.0.2

# Translations
django-modeltranslation==0.12

# GraphQL
django-filter==1.0.1
graphene==1.1.3
graphene-django==1.2.1
